+++ src/dm_preprocessor/tests/mod.rs
@@ -1,3 +1,1 @@
-mod condense_lines;
-mod should_start_new_token;
 mod token_handling;
